+++ components/tour/useMergedType.ts
